pskiplist: add standalone-repo readiness (license, docs, CI, sanitizers, lint)
Prepares pskiplist to be carved out into its own repository: MIT LICENSE,
README, a header-only usage example, a CHANGELOG, GitHub Actions CI (build
+ test on clang/gcc, plus ASan+UBSan and TSan matrix jobs), and
clang-format/clang-tidy configs (readability-* and modernize-* included,
scoped to examples/basic_usage.cpp since it alone drives the full public
API surface).

+++ implementation/pskiplist/tests/normal/test_checkpoint.cpp
@@ -89,7 +89,8 @@ TEST_CASE("concurrent put/remove/checkpoint from multiple threads stays consiste
   constexpr int kSharedKeys = 8;
   constexpr int kIterationsPerThread = 1500;
   TempFile tmp("checkpoint_concurrent");
-  PSkipList<int> skiplist(tmp.path(), capacity_bytes_for_nodes<int>(kThreads * kIterationsPerThread));
+  PSkipList<int> skiplist(tmp.path(),
+                          capacity_bytes_for_nodes<int>(static_cast<size_t>(kThreads) * kIterationsPerThread));
 
   std::atomic<bool> stop{false};
   std::thread checkpointer([&] {
